backend/requirements.txt
Django~=4.2
psycopg2~=2.9

gunicorn~=21.2
requests~=2.31
djangorestframework~=3.14
drf-extra-fields~=3.7
django-filter~=23.4
